Restarting project!
It has been a long while, thought I would try my hand again at doing the
Blackjack project!

Going to redo this using Python and the Click library.

I know that there is a new Click-based library that has better
formatting, but I can't find it within 2 minutes of Googling, so we will
go wtih this for now.