Committing the base of our project.
I have not even started on any of the underlying blackjack game itself,
I have only been working on getting the tooling and stuff set up so far.
We have the basics down, and now we need to start working on the very
first parts of playing the game.

We will then also need to add tests once we get some stuff done! You can
do test-driven development, but sometimes that's not feasible. Will
double-check if it is here and advocate if I can.