Finishing up 1.0 of Deck.
This is required to really work on blackjack, so I wanted to get it out
of the way first.

I think I have everything I will need for now, but there may be some
nice-to-have's that I will want to add in the future.

Have not done testing yet, so this may still change slightly. However,
the methods are present that I believe I will need.